Move `Transform` to dedicated module
